ST08-02 | Fix php stan and eslint

See merge request e16/courses/software-development/ss23/group01!46

+++ backend/database/seeders/DatabaseSeeder.php
@@ -31,16 +31,16 @@ class DatabaseSeeder extends Seeder
         $mapVertices = Map::vertices();
         /** @var Vertex $vertex */
         foreach ($mapVertices as $vertex) {
-            $id = (int)Str::remove(search: 'n', subject: (string)$vertex->getId());
+            $id = Str::remove(search: 'n', subject: (string)$vertex->getId());
             if (MapVertex::find($id) === null) {
                 MapVertex::factory(1)->create(['id' => $id]);
             }
         }
 
         if (TransportRequest::all()->count() === 0) {
-            /** @var User $user */
             $userBar = $output->createProgressBar(count(User::all()));
             $userBar->start();
+            /** @var User $user */
             foreach (User::all() as $user) {
                 if ($user->isAuctioneer()) {
                     continue;
@@ -63,7 +63,11 @@ class DatabaseSeeder extends Seeder
                         'origin_node' => $origin_node_id,
                         'destination_node' => $destination_node_id,
                     ]);
-                    if ($this->vehicleRoutingService->hasOptimalPath(array_merge($transportRequests, [$transportRequest]))) {
+                    if (
+                        $this->vehicleRoutingService->hasOptimalPath(
+                            array_merge($transportRequests, [$transportRequest])
+                        )
+                    ) {
                         $transportRequests[] = $transportRequest;
                         $output->info('Got feasible tr, adding');
                         $trBar->advance();
